requirements.txt
AudioConverter==1.0.0
audioread==3.0.0
autograd==1.6.2
av==10.0.0
chumpy==0.70
easydict==1.10
einops==0.6.1
ftfy==6.1.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
librosa==0.10.0.post2
matplotlib==3.4.3
numpy==1.23.0
Pillow==9.5.0
pyrender==0.1.45
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
smplx==0.1.28
tensorboard==2.13.0
torch==1.12.0
torch-tb-profiler==0.4.1
torchaudio==0.12.1+cu113
torchvision==0.13.0
tqdm==4.65.0
trimesh==3.22.1